pyproject.toml
[project]
name = "consilium-magi"
version = "0.1.0"
description = "MAGI deliberation system — three AI agents debate any topic and vote"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "groq>=0.9.0",
    "python-dotenv>=1.0.0",
]
